Fixed Memory Error with reading WAV data -all samples are now of type float -only add certain ratio of samples to the List -calculate max sample value depending on the bitrate

+++ LiveDescribe/LiveDescribe/View Model/VideoControl.cs
@@ -14,7 +14,8 @@ namespace LiveDescribe.View_Model
        // private Video _video;
         private ILiveDescribePlayer _mediaVideo;
         private AudioUtility _audioOperator;
-        private List<double> _waveFormData;
+        //private List<double> _waveFormData;
+        private List<float> _waveFormData;
        // private MediaElement _videoMedia;
       //  private DispatcherTimer _videoTimer;
         #endregion
